[BranchRelaxation] Remove quadratic behavior in relaxation
Currently, we recompute block offsets after each relaxation. This causes
the complexity to be O(n^2) in the number of instructions, inflating compile
time.

If we instead recompute block offsets after each iteration of the outer
loop, the complexity is O(n). Recomputing offsets in the outer loop will
cause some out-of-range branches to be missed in the inner loop, but
they will be relaxed in the next iteration of the outer loop.

This change may alter the order that branches are relaxed, changing the
order of trampolines at the end of the function.

This change may also  introduce unnecessary relaxations for an
architecture where the relaxed branch is smaller than the unrelaxed
branch, but AFAIK there is no such architecture.