Reapply "[llvm/DWARF] Recursively resolve DW_AT_signature references" (#99444)
The previous version introduced a bug (caught by cross-project tests).
Explicit signature resolution is still necessary when one wants to
access the children (not attributes) of a given DIE.

In the new version of the patch, I keep the `resolveTypeUnitReference`
function, and call it in the function doing child iteration.

I've also extended the prettyprint_type_units test case to cover this
scenario (and also reduced it to remove irrelevant details).